move HTTPResponse.refresh into netlib
